fix: Mark `dist/api` as the main module instead of `dist/plugin`
BREAKING CHANGE: `dist/api` is the main module in `package.json` instead of `dist/plugin`.

The plugin was not meant to import. The API contained named exports
to be consumed. If you did import the plugin, use the full path
`requirejs-esm/dist/plugin`, just like it was used in the RequireJS
configuration of the `esm` path alias.